Added UnicodeConverter.sh by Hagb
Added bash script version for use in Linux, the function is same as the python and executable, which is to convert text to the unicode escape. After setting a hot key to run this script, you can use the hot key for a input box and then input the original text for send.